refactor: Trip Entity 분리
- 기존에 Post 도메인 안에 있던 Trip을 따로 분리함
- TripController
- CreateOrUpdateTripRequest
- GetTripResponse
- TripResponse
- Trip
- TripRepository
- TripService
