+++ src/main/java/com/triprint/backend/domain/trip/entity/Trip.java
@@ -3,7 +3,7 @@
 // (powered by FernFlower decompiler)
 //
 
-package com.triprint.backend.domain.post.entity;
+package com.triprint.backend.domain.trip.entity;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
@@ -20,6 +20,7 @@ import javax.persistence.OneToMany;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
+import com.triprint.backend.domain.post.entity.Post;
 import com.triprint.backend.domain.user.entity.User;
 
 import lombok.Builder;
