+++ src/main/java/com/triprint/backend/domain/trip/repository/TripRepository.java
@@ -1,8 +1,8 @@
-package com.triprint.backend.domain.post.repository;
+package com.triprint.backend.domain.trip.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.triprint.backend.domain.post.entity.Trip;
+import com.triprint.backend.domain.trip.entity.Trip;
 
 public interface TripRepository extends JpaRepository<Trip, Long> {
 }
